# Improve HashGenerator.getPartition performance