Port of https://github.com/prestodb/presto/pull/13611.
Change the mod operation to bit shifting in HashGenerator.getPartition.
BenchmarkPartitionedOutputOperator shows ~10% gain.

Before
```
(channelCount)  (nullRate)  (partitionCount)  (positionCount)  (type)  Mode  Cnt     Score    Error  Units
             1           0               256             8192  BIGINT  avgt   10  2300.677 ± 49.181  ms/op
             1         0.2               256             8192  BIGINT  avgt   10  2292.622 ± 60.765  ms/op
             2           0               256             8192  BIGINT  avgt   10  3411.225 ± 47.299  ms/op
             2         0.2               256             8192  BIGINT  avgt   10  3601.043 ± 49.378  ms/op
```

After
```
(channelCount)  (nullRate)  (partitionCount)  (positionCount)  (type)  Mode  Cnt     Score     Error  Units
             1           0               256             8192  BIGINT  avgt   10  1962.333 ±  36.066  ms/op
             1         0.2               256             8192  BIGINT  avgt   10  2067.068 ± 137.070  ms/op
             2           0               256             8192  BIGINT  avgt   10  3070.689 ±  41.821  ms/op
             2         0.2               256             8192  BIGINT  avgt   10  3329.175 ± 122.881  ms/op
```