@lukasz-stec 2021-09-06
> > Port of
> 
> please add `Extracted-From:` to the commit message, as eg in [d56a6d9](https://github.com/trinodb/trino/commit/d56a6d9e36547d577865ef5ec764e1cb60d72ffb)

done