@pettyjamesm 2021-09-06
A side note about this “fast range” bucketing (more or less, this algorithm: https://github.com/lemire/fastrange)- it depends on having a good hash function produce the input before bucketing, because low bits are essentially discarded when determining the bucket (unlike with modulo). 

Types that currently pass their bit patterns through directly as their hash code ([like ShortDecimal](https://github.com/trinodb/trino/blob/040d857e9a2b15c4efa80a306b544392e56e0da6/core/trino-spi/src/main/java/io/trino/spi/type/ShortDecimalType.java#L128,L131)) will all fall into bucket 0 for all but extremely large values, and you’ll see a huge problem with data skew.