@sopel39 2021-09-07
@pettyjamesm thanks. Addressed in https://github.com/trinodb/trino/pull/9138